code updates
updated docstrings
